refactor: core 디렉토리 zzapdux index에서 export
