+++ core/index.js
@@ -1,1 +1,2 @@
 export * from "./db.js";
+export * from "./zzapdux.js";
