+++ store/index.js
@@ -1,4 +1,4 @@
-import { createStore } from "../core/zzapdux.js";
+import { createStore } from "../core/index.js";
 import { reducer } from "./reducer.js";
 
 export const store = createStore(reducer);
